The code is now able to handle more than 2^32 triangles (as result of multiple meshes and/or instances, not in a single mesh)

+++ include/luxrays/core/extmeshcache.h
@@ -41,7 +41,7 @@ public:
 	void SetDeleteMeshData(const bool v) { deleteMeshData = v; }
 
 	void DefineExtMesh(const std::string &fileName,
-		const long plyNbVerts, const long plyNbTris,
+		const u_int plyNbVerts, const u_int plyNbTris,
 		Point *p, Triangle *vi, Normal *n, UV *uv,
 		luxrays::Spectrum *cols, float *alphas,
 		const bool usePlyNormals);
